requirements.txt
streamlit==1.37.0
youtube-transcript-api==0.6.2
yt-dlp==2024.8.6
huggingface_hub==0.24.6
sentence-transformers==3.0.0
numpy>=1.26.4
faiss-cpu>=1.8.0.post1
pytest==8.2.2
black==24.4.2
ruff==0.5.0
mypy==1.10.0
python-dotenv==1.0.1
